Add raw print test module and integrate printing
